В компонент Scroll добавлен необязательный пропс, позволяющий скролить захватывая элемент. Компонент Наши клиенты о нас переделан на использование Scroll
