+++ src/widgets/ReviewsBlock/ui/ReviewsBlock/ReviewsBlock.tsx
@@ -1,4 +1,4 @@
-import { MouseEvent, useEffect, useRef, type FC } from 'react'
+import { useEffect, type FC } from 'react'
 import { useSelector } from 'react-redux'
 
 import { StateSchema } from '@/app/providers/StoreProvider'
@@ -10,6 +10,7 @@ import { IFeedback } from '@/features/Reviews/model/types/types'
 import { useAppDispatch } from '@/shared/libs/hooks/store'
 import Heading, { HeadingType } from '@/shared/ui/Heading/Heading'
 import Link from '@/shared/ui/Link/Link'
+import Scroll from '@/shared/ui/Scroll/Scroll'
 
 import styles from './reviewsBlock.module.scss'
 
@@ -31,36 +32,12 @@ const ReviewsBlock: FC<Props> = props => {
 
   const dispatch = useAppDispatch()
   const reviews = useSelector((store: StateSchema) => store.feedbacks)
-  const listRef = useRef<HTMLDivElement>(null)
-  const startX = useRef(0)
-  const startScrollLeft = useRef(0)
-  const isDragging = useRef(false)
 
   useEffect(() => {
     dispatch(getFirstFeedbacks())
 
     dispatch(getAverageMark())
   }, [])
-
-  const handleMouseDown = (e: MouseEvent<HTMLDivElement, globalThis.MouseEvent>) => {
-    startX.current = e.clientX
-    if (listRef.current) {
-      startScrollLeft.current = listRef.current.scrollLeft
-    }
-    isDragging.current = true
-  }
-
-  const handleMouseMove = (e: MouseEvent<HTMLDivElement, globalThis.MouseEvent>) => {
-    if (!isDragging.current) return
-    const deltaX = e.clientX - startX.current
-    if (listRef.current) {
-      listRef.current.scrollLeft = startScrollLeft.current - deltaX
-    }
-  }
-
-  const handleMouseUp = () => {
-    isDragging.current = false
-  }
 
   return (
     <section className={styles.wrapper}>
@@ -75,13 +52,7 @@ const ReviewsBlock: FC<Props> = props => {
           {IconLink({ styles: styles.svg })}
         </Link>
       </article>
-      <div
-        ref={listRef}
-        onMouseMove={handleMouseMove}
-        onMouseDown={handleMouseDown}
-        onMouseUp={handleMouseUp}
-        onMouseLeave={handleMouseUp}
-        className={styles.list}>
+      <Scroll className={styles.list} withManualGrip={true}>
         <ReviewCard
           key="feadbackHeader"
           pk={0}
@@ -102,7 +73,7 @@ const ReviewsBlock: FC<Props> = props => {
             index={index}
           />
         ))}
-      </div>
+      </Scroll>
     </section>
   )
 }
