+++ src/widgets/ViewedProducts/ViewedProducts.module.scss
@@ -33,5 +33,6 @@
 
   &__scroll {
     width: 100%;
+    user-select: none;
   }
 }
